Fix broken input source in string manipulation funcs
These funcs seem to fail when resolving the input (stdin, args)
hanging for ever when called with an empty string as argument.

```sh
val="$(trim "")" # causes prompt to hang for ever
```

Checking if the stdin has any input data can work as a
conditional safe gate to catch calls with empty string args.