Fix broken pipeline flow after a script exits with non-zero
It turns out that the || in the pipeline after unmounting partitions
causes the flow to break in cases a script exits with non-zero code.
So when the user wants to exit the installation process the || pipeline
breaks the flow and still continues to execute the following commands.
Considering this echo is not so critical we decided to remove it.