Have installation files into the right path during installation
Installation file have to be saved into a path which should be accessible
by the bash session a script is taking control within. For instance the
system installation script should find the installation scripts under the
root folder, in the same way the desktop installation script should expect
to access installation files under the user's home folder as thoses process
meant to be only for user tasks.