Fix broken show status after setting LANG locale variable
The show status has been considering only the first lines of the
output the localectl was giving, that is not the case anymore. The
reason is that setting for instance the LANG locale var other than
the rest locale vars make the localectl status prints all the other
variables after the LANG variable, so breaking the show stauts layout.