Make sure sed iteration stops at the next output line
For outputs which are not connected no resolution data are provided
so when a request is comming for such an output, the rate extraction
process is considering as input the next output's resolution data.
A disconnected outout is not expected to have a rate, so we have to
stop early at the next output's head line.