Fix broken extraction of the output rate value
This is tricky due to the way the resolutions and rates are printed
by xrandr, when the current res and rate are in the first line of modes
the rate is extracted correctly. In the case in which the output has set
to a res deep down in the modes lines the extraction of rate breaks. This
is happening because the sed doesn't stop at the first * match, which should
be the expected behavior. In addition for disconnected and inactive outputs
the rate extracted is not N/A as it should be but the rate of the next output
if there is an active one.