Fix and improve validation when user matches outputs to layout spots
By this change we replace the broken sum of indices mechanism to ensure
that the user input has unique values and each them within the expected
range. In addition more checks are added to validate inputs that include
zerofill numbers like 0 and 00 which should be considered as duplicates.

Long story short we first clean extra spaces then check if the input comes
as a list of non-zerofill ints (space separated), then remove dups and assert
if the values remain are within the expected index reange and are in total
equal to the length of outputs.