Do not include connected and inactive monitors in layout signatures
Having the layout signature built on any connected outputs active or inactive
causes the restore layout operation to break. The reason is that the restore
code is expecting to match that signature built at a previous save layout operation
which means that in order to restore it, any connected and inactive monitor still
have to be hooked into the same output port, otherwise it will not match the current
mapping of active monitors.

This change fixes this malfunction by using only the active monitors in the signature.