Refactor prompt input in tools to not ignore args given as empty string
In loop mode the user input should now passed as it is through an eval command
in order to retain any given arguments even those passed as empty string like
so, `command arg1 "" arg2`. In this example the command function which is expecting
3 args will take exactly 3 arguments one of which will be empty. Previously the second
argument was always being ignored. That way we can now trigger user input on not
given arguments without breaking the positional order of other args.