Fix desktop workspaces after various display layout events
It turns out that operations like set, restore and fix layout have an impact
on the current workspaces map, anytime xrandr modifies the monitors layout
the workspaces map should correct any left inconsistencies. This is also
when a new monitor is added by the user via the 'displays set on' cmd.