Fix broken swap desktops binding for focused desktops
It turns out that the scratchpad causes the bspc to reject any
calls to swap desktops. The root of the cause is that bspc is
checking for sticky nodes in the curretly focused desktop and
if finds any cancels the swap operation. So by removing the
sticky mode from the scratchpad we solve this issue.