Drop ipc modules for audio devices due to high cpu usage
It turns out that when the user is key down pressing the volume up/down
shortcuts of either the output or the input the usage of cpu is bumping
high due to the load of consequitives polybar-msg calls. Eventhough the
ipc practice seemed to be the only solution, we drop this feature and
falling back to the internal pulseaudio module for now discarding the
microphone module. Perhaps the PR #2238 will be soon merged in the polybar
repo.