Make sure the logger resolves the correct log file path
Each time the installer moves (via chroot) to the new system (/mnt), in order
to install the system, desktop, stack and tools packages, it loses access
to the original log file which is initially located under the /var/log
directory of the archiso file system. To anticipate that problem we move
the log file under the /mnt/var/log folder after the bootstrap step completes
so the afformentioned script can now access the log file under their file system (/mnt).

Unfortunately all this, having the log file under the new system (/mnt), causes
another issue, as the install script is now unable to log correctly because this
script runs under the archiso file system (not chroot) and so it has to resolve
the correct log file path which has been changed to /mnt.

We assume that when the install script tries to log into the log file and find no
such file this file should be located under the new system fs.