Restore func names back to their previous name (refs #2e56938)
This commit restores back the names of the funcs have been changed in previous
commit (#2e56938) in order to fix func name overrides among funcs of the tools
scripts. After testing only for execution paths there is no need for a func name
to be globally unique unless it is crossing with a script file on the same execution
path via `source` cmds.

Refs #2e56938