Hard code /usr/local/stack PATH variable in any systemd user services context
This commit fixes:
Due to the reason the systemd services being executed out of the user's session
context any custom PATH set by bashrc or profile getrs excluded from the PATH
variable, so ending with tools global binaries not be available. Hard coding
the path as environment variable we solve this issue.