Fix command prompt overlapping on the same line typing long commands
This commit fixes:
When typing long command the command prompt is overlapped by the same line,
this caused by the not printable special chars length is longer than the actual
PS1 printable section. That's what \[ and \] in bash are there for: they
guard the contained characters from being counted towards the prompt length.